+++ packages/ui/src/components/Roving/RovingTabIndexRoot.tsx
@@ -1,44 +1,16 @@
-import isHotkey from "is-hotkey"
 import {
   ReactNode,
   useCallback,
   useRef,
   useState,
-  FocusEvent,
-  MouseEvent,
-  KeyboardEvent,
-  ComponentPropsWithoutRef,
-  ElementType,
-  MutableRefObject,
+  type ComponentPropsWithoutRef,
+  type ElementType,
 } from "react"
 import { useControlledState } from "../../hooks/useControllableState"
-import { createContext } from "../../hooks/createContext"
+import { RovingTabIndexProvider } from "./useRovingTabIndex"
 import { composeRefs } from "../../hooks/useComposedRefs"
 
-export type RovingTabIndexItem = {
-  value: string
-  element: HTMLElement
-}
-
-function onFocusFirst(candidates: HTMLElement[]) {
-  const previousFocus = document.activeElement
-  while (document.activeElement === previousFocus && candidates.length > 0) {
-    candidates.shift()?.focus()
-  }
-}
-
-type RovingTabIndexContext = {
-  currentFocusedValue: string | null
-  setFocusableId: (id: string) => void
-  onShiftTab: () => void
-  getOrderedItems: () => RovingTabIndexItem[]
-  elements: MutableRefObject<Map<string, HTMLElement>>
-}
-
-const [RovingTabIndexProvider, useRovingTabIndexContext] =
-  createContext<RovingTabIndexContext>("rovingTabIndex")
-
-const NODE_SELECTOR = "data-roving-tabindex-node"
+export const NODE_SELECTOR = "data-roving-tabindex-node"
 const ROOT_SELECTOR = "data-roving-tabindex-root"
 const NOT_FOCUSABLE_SELECTOR = "data-roving-tabindex-not-focusable"
 const SELECTOR_ACTIVE = `:where([${NODE_SELECTOR}=true]):not(:where([${NOT_FOCUSABLE_SELECTOR}=true] *))`
@@ -48,11 +20,19 @@ type RovingTabIndexRootBaseProps<T> = {
   active?: string
   setActive?: (value: string) => void
   as?: T
+  dir?: "horizontal" | "vertical"
 }
 
 type RovingTabIndexRootProps<T extends ElementType> =
   RovingTabIndexRootBaseProps<T> &
     Omit<ComponentPropsWithoutRef<T>, keyof RovingTabIndexRootBaseProps<T>>
+
+function onFocusFirst(candidates: HTMLElement[]) {
+  const previousFocus = document.activeElement
+  while (document.activeElement === previousFocus && candidates.length > 0) {
+    candidates.shift()?.focus()
+  }
+}
 
 export const RovingTabIndexRoot = <T extends ElementType>({
   children,
@@ -60,6 +40,7 @@ export const RovingTabIndexRoot = <T extends ElementType>({
   setActive,
   as,
   ref,
+  dir = "horizontal",
   ...props
 }: RovingTabIndexRootProps<T>) => {
   const Component = as || "div"
@@ -69,7 +50,6 @@ export const RovingTabIndexRoot = <T extends ElementType>({
     prop: active,
     onChange: setActive,
   })
-
   const rootRef = useRef<HTMLDivElement | null>(null)
   const elementsRef = useRef<Map<string, HTMLElement>>(new Map())
   const getOrderedItems = useCallback(() => {
@@ -96,17 +76,16 @@ export const RovingTabIndexRoot = <T extends ElementType>({
       currentFocusedValue={value}
       getOrderedItems={getOrderedItems}
       elements={elementsRef}
+      dir={dir}
     >
       <Component
         {...{ [ROOT_SELECTOR]: true }}
-        // tabIndex={isShiftTabbing ? -1 : 0}
+        tabIndex={isShiftTabbing ? -1 : 0}
         onFocus={(e) => {
-          console.log("parent focus")
           if (e.target !== e.currentTarget) return
           if (isShiftTabbing) return
           const orderedItems = getOrderedItems()
           if (orderedItems.length === 0) return
-
           const candidates = [
             elementsRef.current.get(value ?? ""),
             ...orderedItems.map((i) => i.element),
@@ -122,51 +101,4 @@ export const RovingTabIndexRoot = <T extends ElementType>({
       </Component>
     </RovingTabIndexProvider>
   )
-}
-
-export const useRovingTabIndex = (value: string) => {
-  const {
-    currentFocusedValue,
-    setFocusableId,
-    onShiftTab,
-    getOrderedItems,
-    elements,
-  } = useRovingTabIndexContext("rovingTabIndex")
-
-  return {
-    getOrderedItems,
-    isFocusable: currentFocusedValue === value,
-    getRovingProps: <T extends ElementType>(
-      props?: ComponentPropsWithoutRef<T>,
-    ) => ({
-      ...props,
-      ref: (element: HTMLElement | null) => {
-        if (element) {
-          elements.current.set(value, element)
-        } else {
-          elements.current.delete(value)
-        }
-      },
-      onMouseDown: (e: MouseEvent) => {
-        props?.onMouseDown?.(e)
-        if (e.target !== e.currentTarget) return
-        setFocusableId(value)
-      },
-      onKeyDown: (e: KeyboardEvent) => {
-        props?.onKeyDown?.(e)
-        if (e.target !== e.currentTarget) return
-        if (isHotkey("shift+tab", e)) {
-          onShiftTab()
-          return
-        }
-      },
-      onFocus: (e: FocusEvent) => {
-        props?.onFocus?.(e)
-        if (e.target !== e.currentTarget) return
-        setFocusableId(value)
-      },
-      [NODE_SELECTOR]: true,
-      tabIndex: currentFocusedValue === value ? 0 : -1,
-    }),
-  }
 }
